Fixed return value

+++ src/RNG.h
@@ -12,9 +12,9 @@ class RNG
         virtual int x() const = 0;
         virtual int y() const = 0;
 
-        virtual uint8_t red() const = 0;
-        virtual uint8_t green() const = 0;
-        virtual uint8_t blue() const = 0;
+        virtual int red() const = 0;
+        virtual int green() const = 0;
+        virtual int blue() const = 0;
 };
 
 #endif
